test(parser-jvm-core): move ClassInfoModel tests to the general test directory
