+++ packages/java/parser-jvm-core/src/test/java/dev/hilla/parser/models/ClassInfoModelBasicTests.java
@@ -1,4 +1,4 @@
-package dev.hilla.parser.models.classinfomodel;
+package dev.hilla.parser.models;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
